Implement user sidebar
On the right, all connected users are shown. This is hidden for screen widths below 500px.
Therefore, a new message type 'connected' is implemented which provides all connected users including their ids for rendering.
User functionality was implemented in the sample project this project is based on, so it only had to be adjusted.