fix missing cookies in dev environment with http
setting {cookie: {secure: true}} in session config will only set a
cookie if the request was done securely, i.e. using https with a custom
cert or via a trusted proxy.

But when developing locally on http://localhost neither may be the case
and the missing cookie will cause a redirect loop with OAuth or broken
connection with manual login.
Therefor set `secure: false` if neither a custom cert nor trusted proxy
is configured.